% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyncNomis-package.R
\docType{package}
\name{AsyncNomis-package}
\alias{AsyncNomis}
\alias{AsyncNomis-package}
\title{AsyncNomis: Shiny implementation for asynchronous interactive querying of Nomis data}
\description{
Asynchronous querying of Nomis data in Shiny.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/konradzdeb/AsyncNomis}
  \item Report bugs at \url{https://github.com/konradzdeb/AsyncNomis/issues}
}

}
\author{
\strong{Maintainer}: Konrad Zdeb \email{konrad.zdeb@me.com}

}
\keyword{internal}
